+++ sdk/metricsadvisor/azure-ai-metricsadvisor/tests/async_tests/base_testcase_async.py
@@ -190,8 +190,8 @@ class TestMetricsAdvisorAdministrationClientBaseAsync(AzureTestCase):
                     sensitivity=50,
                     anomaly_detector_direction="Both",
                     suppress_condition=SuppressCondition(
-                        min_number=50,
-                        min_ratio=50
+                        min_number=5,
+                        min_ratio=5
                     )
                 )
             )
@@ -305,13 +305,13 @@ class TestMetricsAdvisorAdministrationClientBaseAsync(AzureTestCase):
             metric_id=data_feed.metric_ids['cost'],
             description="My test metric anomaly detection configuration",
             whole_series_detection_condition=MetricDetectionCondition(
-                cross_conditions_operator="AND",
+                condition_operator="AND",
                 smart_detection_condition=SmartDetectionCondition(
                     sensitivity=50,
                     anomaly_detector_direction="Both",
                     suppress_condition=SuppressCondition(
-                        min_number=50,
-                        min_ratio=50
+                        min_number=5,
+                        min_ratio=5
                     )
                 ),
                 hard_threshold_condition=HardThresholdCondition(