+++ sdk/sql/azure-mgmt-sql/azure/mgmt/sql/operations/_servers_operations.py
@@ -71,7 +71,7 @@ class ServersOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-11-01-preview"
+        api_version = "2021-02-01-preview"
         accept = "application/json"
 
         def prepare_request(next_link=None):
@@ -151,7 +151,7 @@ class ServersOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-11-01-preview"
+        api_version = "2021-02-01-preview"
         accept = "application/json"
 
         # Construct URL
@@ -202,7 +202,7 @@ class ServersOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-11-01-preview"
+        api_version = "2021-02-01-preview"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
@@ -333,7 +333,7 @@ class ServersOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-11-01-preview"
+        api_version = "2021-02-01-preview"
 
         # Construct URL
         url = self._delete_initial.metadata['url']  # type: ignore
@@ -443,7 +443,7 @@ class ServersOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-11-01-preview"
+        api_version = "2021-02-01-preview"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
@@ -579,7 +579,7 @@ class ServersOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-11-01-preview"
+        api_version = "2021-02-01-preview"
         accept = "application/json"
 
         def prepare_request(next_link=None):
@@ -644,7 +644,7 @@ class ServersOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-11-01-preview"
+        api_version = "2021-02-01-preview"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
@@ -780,7 +780,7 @@ class ServersOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-11-01-preview"
+        api_version = "2021-02-01-preview"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"