+++ sdk/textanalytics/azure-ai-textanalytics/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 5.1.0 (2021-07-08)
+## 5.1.0 (2021-07-07)
 
 This version of the SDK defaults to the latest supported API version, which currently is `v3.1`.
 Includes all changes from `5.1.0b1` to `5.1.0b7`.